(1248) Split up userModel
