+++ packages/api/server/models/userModel/_common/getNationalAdmins.js
@@ -1,0 +1,16 @@
+const query = require('./query');
+
+module.exports = () => query([
+    {
+        fk_role: ['national_admin'],
+    },
+    {
+        fk_status: ['active'],
+    },
+    {
+        organization_active: {
+            query: 'organizations.active',
+            value: [true],
+        },
+    },
+], {});
