+++ packages/api/server/models/userModel/deactivate.js
@@ -1,0 +1,16 @@
+const { sequelize } = require('#db/models');
+
+module.exports = id => sequelize.query(
+    `UPDATE
+        users
+    SET
+        fk_status = 'inactive'
+    WHERE
+        user_id = :id
+    `,
+    {
+        replacements: {
+            id,
+        },
+    },
+);
