+++ packages/api/server/models/userModel/delete.js
@@ -1,0 +1,10 @@
+const { sequelize } = require('#db/models');
+
+module.exports = id => sequelize.query(
+    'DELETE FROM users WHERE users.user_id = :id',
+    {
+        replacements: {
+            id,
+        },
+    },
+);
