+++ packages/api/server/models/userModel/findAll.js
@@ -1,0 +1,3 @@
+const query = require('./_common/query');
+
+module.exports = user => query([], { auth: false, extended: false }, user, 'list');
