+++ packages/api/server/models/userModel/findByIds.js
@@ -1,0 +1,12 @@
+const query = require('./_common/query');
+
+module.exports = (user, userIds) => query(
+    [
+        {
+            user_id: userIds,
+        },
+    ],
+    { auth: false, extended: false },
+    user,
+    'list',
+);
