+++ packages/api/server/models/userModel/findByOrganization.js
@@ -1,0 +1,13 @@
+const query = require('./_common/query');
+
+module.exports = (organizationId, filters = {}) => query(
+    [
+        {
+            organization: {
+                query: 'organizations.organization_id',
+                value: [organizationId],
+            },
+        },
+    ],
+    filters,
+);
