+++ packages/api/server/models/userModel/findDepartementSummarySubscribers.js
@@ -1,0 +1,31 @@
+const query = require('./_common/query');
+
+module.exports = async () => {
+    const users = await query(
+        [
+            {
+                fk_status: ['active'],
+            },
+            {
+                subscribed_to_summary: true,
+            },
+        ],
+        {},
+    );
+
+    return users.reduce((argAcc, user) => {
+        const acc = { ...argAcc };
+
+        let locationType = user.organization.location.type;
+        if (!acc[locationType]) {
+            locationType = 'departement';
+        }
+
+        acc[locationType].push(user);
+        return acc;
+    }, {
+        nation: [],
+        region: [],
+        departement: [],
+    });
+};
