+++ packages/api/server/models/userModel/upgradeLocalAdmin.js
@@ -1,0 +1,12 @@
+const { sequelize } = require('#db/models');
+
+module.exports = async (userId) => {
+    await sequelize.query(
+        'UPDATE USERS SET fk_role = \'local_admin\' where user_id = :userId',
+        {
+            replacements: {
+                userId,
+            },
+        },
+    );
+};
